% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomFieldGeneration.R
\name{ArbCovNoise}
\alias{ArbCovNoise}
\title{Creates sample paths from a 1D random field having an arbitrary covariance function. The standard is given by the square exponential function: C(h)=exp( -h^2/(4*nu^2) )}
\usage{
ArbCovNoise(N, x = seq(0, 1, length.out = 100), sigma = function(x) {     1
  }, CovFun = function(x, y) {     exp(-(x - y)^2/4/4^2) })
}
\arguments{
\item{N}{Integer amount of realisations of the random field to be generated.}

\item{x}{Vector locations at which the random field is evaluated.}

\item{sigma}{Function computing the pointwise variance of the field. Default value is unit variance everywhere.}

\item{CovFun}{Function taking to arguments, which computes the desired covariance.}
}
\value{
Matrix containing the realisations of the random field as columns. The rows are the locations.
}
\description{
Creates sample paths from a 1D random field having an arbitrary covariance function. The standard is given by the square exponential function: C(h)=exp( -h^2/(4*nu^2) )
}
