% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Miscellaneous.R
\name{bias_Gaussian}
\alias{bias_Gaussian}
\title{This functions computes the bias of certain transformations under Gaussianity}
\usage{
bias_Gaussian(transformation, N, hatd = NULL)
}
\arguments{
\item{transformation}{The considered transformation of the moments}

\item{N}{Vector grid for evaluation of the smoothed field}
}
\value{
Standard error under the assumption the data is Gaussian
}
\description{
This functions computes the bias of certain transformations under Gaussianity
}
