% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoveringRateSimulations.R
\name{covRate_2sample}
\alias{covRate_2sample}
\title{Computing the covering rate of a given method for constructing simultaneous confidance bands for the mean for a given error process.}
\usage{
covRate_2sample(trials, N1 = 10, N2 = 10, level = 0.95, x = seq(0, 1,
  length.out = 100), mu1 = function(x) {     sin(8 * pi * x) * exp(-3 * x) },
  mu2 = function(x) {     sin(8 * pi * x) * exp(-3 * x) },
  noise1 = BernsteinSumNoise, noise2 = BernsteinSumNoise,
  sigma1 = function(x) {     ((0.6 - x)^2 + 1)/6 }, sigma2 = function(x) {  
    ((0.6 - x)^2 + 1)/6 }, sd_ObsNoise1 = 0, sd_ObsNoise2 = 0,
  method = "tGKF", param_method = NULL, ...)
}
\arguments{
\item{trials}{Integer amount of trials to obtain the covering rate.}

\item{N1}{Integer sample size for sample 1.}

\item{N2}{Integer sample size for sample 2.}

\item{level}{Numeric the targeted covering probability. Must be strictly between 0 and 1.}

\item{method}{String the name of the method used for quantile estimation. Possibilities are "tGKF", "GKF". Default value is "tGKF".}

\item{Mboots}{Numeric the number of bootstrap replicats used, if the quantile is estimated using a bootstrap method. Default value is 5000.}

\item{noise}{Function a function generating random realizations.}

\item{mean}{Vector a vector containing the values of the true mean of the random functions.}
}
\value{
list with elements
 \itemize{
  \item cov.rate Numeric simulated covering rate
  \item trials Integer number of simulations to estimate the covering rate
  \item BootsReplicats Integer number of bootstrap replicates
  \item nSamp Integer number of
}
}
\description{
Computing the covering rate of a given method for constructing simultaneous confidance bands for the mean for a given error process.
}
