% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoveringRateSimulations.R
\name{covRate_simulation}
\alias{covRate_simulation}
\title{Simulates the covering rate of different methods for constructiong of simultaneous confidence bands (SCBs) for the mean of functional signal plus noise model (SpN) with different error processes.}
\usage{
covRate_simulation(trials, scenario, param_scenario = NULL, method,
  param_method = NULL, level = 0.95, N, mu = function(x) {     rep(0,
  length(x)) }, noise = GaussDensitySumNoise, sigma = function(x) {    
  rep(1, length(x)) }, sd_ObsNoise = 0, x = seq(0, 1, length.out =
  100), ...)
}
\arguments{
\item{trials}{Integer number of trials to estimate the covering rate.}

\item{scenario}{String specifying the scenario. Options are:
\itemize{
  \item SpN:   Signal plus Noise Model with one or two dimensional domain. Make sure that the functions 'mu', 'sigma' and 'noise' produce a correct and consistent output for the dimension you want to simulate.
  \item Scale1d: Scale field of Signal plus Noise Model with one dimensional domain
}}

\item{param_scenario}{list containing parameters for the scenario.
\itemize{
  \item SpN:     'SmoothWeights': Either NULL or array of size Keval x length(x).
  \item Scale1d: Scale field of Signal plus Noise Model with one dimensional domain
}}

\item{method}{String the method to compute the SCBs. Options are "tGKF", "GKF", "Bootstrap", "Bootstrapt", "MultiplierBootstrap", "MultiplierBootstrapt". Default value is "tGKF".}

\item{level}{Numeric the targeted covering probability. Must be strictly between 0 and 1. Default is 0.95.}

\item{N}{Integer number of samples used to compute the SCBs.}

\item{mu}{Function computing the population mean of the SpN.}

\item{noise}{Function generating random N realizations of a noise process. The output must be an array of dimension K1 x ... x Kd x N.}

\item{sigma}{Function defining the pointwise variance in the SpN.}

\item{sd_ObsNoise}{Numeric standard deviation of zero mean Gaussian observation noise, which will be added to the SpNM. Default value is 0.}

\item{x}{Vector of length K containing the grid points, on which the simulated processes are evaluated.}

\item{...}{Additional parameter for 'FunctionalDataSample()', i.e. parameter for the noise process.}
}
\value{
list with elements
 \itemize{
  \item cov.rate Number (simulated covering rate)
  \item trials
  \item BootsReplicats
  \item nSamp
}
}
\description{
Simulates the covering rate of different methods for constructiong of simultaneous confidence bands (SCBs) for the mean of functional signal plus noise model (SpN) with different error processes.
}
