% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomFieldGeneration.R
\name{DegrasNonGaussNoise}
\alias{DegrasNonGaussNoise}
\title{Creates sample paths from a 1D non Gaussian field as used in Degras (2011, Simultaneous confidence bands for nonparametric regression with functional data. Statistica Sinica, 1735-1765.).}
\usage{
DegrasNonGaussNoise(N, x = seq(0, 1, length.out = 100), sigma = function(x)
  {     1 })
}
\arguments{
\item{N}{Integer amount of realisations of the random field to be generated.}

\item{x}{Vector locations at which the random field is evaluated.}

\item{Function}{computing the pointwise variance of the field. Default value is unit variance everywhere.}
}
\value{
Matrix containing the realisations of the random field as columns. The rows are the locations.
}
\description{
Creates sample paths from a 1D non Gaussian field as used in Degras (2011, Simultaneous confidence bands for nonparametric regression with functional data. Statistica Sinica, 1735-1765.).
}
