% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Residuals.R
\name{DeltaMomentResiduals}
\alias{DeltaMomentResiduals}
\title{Computes functional delta residuals for moment based statistics}
\usage{
DeltaMomentResiduals(Y, transformation, moments = NULL, oracle = NULL)
}
\arguments{
\item{Y}{array of dimension K_1 x ... x K_D x N containing N-realizations of
a random field over a D-dimensional domain.}

\item{transformation}{expression of the form f ~ f(mu1, ..., muk ) giving
the transformation of the moments. Can be a "linear", "cohensd", "skewness" or
"kurtosis". Then standard estimators are used.
Current options are "skewness" and "kurtosis".}

\item{moments}{string of the form c("mu1", ..., "muk") containing the names of
the moments contained in the order of transformation input.}
}
\value{
an array representing the delta residuals
}
\description{
Computes functional delta residuals for moment based statistics
}
