% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GaussianKinematicFormulas.R
\name{ECdensity}
\alias{ECdensity}
\title{Computes the d-th EC density of different fields. Currently only "t" is implemented.}
\usage{
ECdensity(x, d, field = "t", df = 1)
}
\arguments{
\item{x}{Numeric location x at which the EC density is to be evaluated}

\item{d}{Numeric d-th EX density}

\item{field}{String type of field for EC densities, currently only "t" (Student-t) and "Gauss" (Gaussian) are supported}

\item{df}{Numeric degree of freedom of t-field}
}
\value{
value of 2d-EC density of a t-field of degree df at location x.
}
\description{
Computes the d-th EC density of different fields. Currently only "t" is implemented.
}
