% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LKCEstimators.R
\name{EulerChar}
\alias{EulerChar}
\title{Estimates the Euler characteristic of the exceedance set A(u)={ s in S | f(s) >u } of a function f:S->R.
4 or 8 connectivity  is currently used and it is implemented up to dimension 2.}
\usage{
EulerChar(f, u, connectivity = 8)
}
\arguments{
\item{f}{Vector/Matrix observation of a random function on a grid}

\item{u}{Vector exceedance values to be computed}

\item{connectivity}{Integer amount of points considered as neighbors}
}
\value{
Vector of Euler characteristics of the exceedance sets for the values u
}
\description{
Estimates the Euler characteristic of the exceedance set A(u)={ s in S | f(s) >u } of a function f:S->R.
4 or 8 connectivity  is currently used and it is implemented up to dimension 2.
}
