% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fairness.R
\name{FairThreshold1D}
\alias{FairThreshold1D}
\title{Estimates from a sample of random functions (for example a bootstrap sample)
a Fair thresholding function q}
\usage{
FairThreshold1D(
  samples,
  x,
  fair,
  fair.type = "linear",
  Splus = rep(TRUE, length(x)),
  Sminus = rep(TRUE, length(x)),
  alpha = 0.95,
  diff.fair = NULL,
  subI = NULL
)
}
\arguments{
\item{x}{vector of length K of locations at which the sample is observed.}

\item{fair}{a vector partitioning the vector x into regions on which the
on which the rejection should be fair. First element must be
x[1] and last x[length(x)]}

\item{alpha}{numeric the targeted upper quantile of the maxiumum of the
absolute value of the random field. Default is 0.95.}

\item{sample}{array of dimension K x N containing N-realizations of
a random field over a 1-dimensional domain.}
}
\value{
list with elements
 \itemize{
  \item q Vector containing the fair piecewise linear thresholding function at each x
  \item qm Vector containing the offset and the slopes of the fair thresholding function
  \item EmpRejections Numeric giving the empirical rejection rate of the fair
  thresholding function with respect to the sample.
}
}
\description{
Estimates from a sample of random functions (for example a bootstrap sample)
a Fair thresholding function q
}
