% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomFieldGeneration.R
\name{FunctionalDataSample}
\alias{FunctionalDataSample}
\title{Function generating a sample of a signal plus noise model in functional data.}
\usage{
FunctionalDataSample(N, x = seq(0, 1, length.out = 100),
  mu = function(x) {     rep(0, length(x)) }, noise = SinCosSumNoise,
  sigma = function(x) {     rep(1, length(x)) }, sd_ObsNoise = 0, ...)
}
\arguments{
\item{N}{Integer amount of realisations of the random field to be generated.}

\item{x}{Vector/Array locations at which the random field is evaluated. Must be a valid input for the functions 'mu', 'noise' and 'sigma'.}

\item{mu}{Function computing the population mean of the signal plus noise model from a vector of locations.}

\item{noise}{Function computing a sample of random fields.}

\item{sigma}{Function computing the mean from a vector.}

\item{sd_ObsNoise}{Numeric positiv real number giving the standard deviation for independent additiv Gaussian observation noise.}

\item{...}{additional parameters for the 'noise' function}
}
\description{
Function generating a sample of a signal plus noise model in functional data.
}
