% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultipleTestProcedures.R
\name{FWE_control}
\alias{FWE_control}
\title{This functions computes the true standard error depending on N for different
functionals of Gaussians}
\usage{
FWE_control(alpha, pvals, correction)
}
\arguments{
\item{alpha}{numeric in (0,1) the overall significance level to be achieved}

\item{pvals}{a vector of pvalues}

\item{correction}{string currently "bonferroni" or "holm" are options}
}
\value{
Standard error under the assumption the data is Gaussian
}
\description{
This functions computes the true standard error depending on N for different
functionals of Gaussians
}
