% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomFieldGeneration.R
\name{GaussDensitySum2DNoise}
\alias{GaussDensitySum2DNoise}
\title{Creates a sample from an 2D random field, which is the sum of 2-D Gaussian densities..}
\usage{
GaussDensitySum2DNoise(N, x = seq(0, 1, length.out = 50),
  sigma = function(x) array(outer(x, x, FUN = function(s, t) (s + 1)/(t^2 +
  1))/3, c(rep(length(x), 2), 1)), randNumber = rnorm, M = 6, bdx = 0.02,
  Fnorm = NULL)
}
\arguments{
\item{N}{Integer amount of realisations of the random field to be generated.}

\item{x}{Vector locations at which the random field is evaluated.}

\item{sigma}{Function computing the pointwise variance of the field. Default value is unit variance everywhere.}

\item{randNumber}{Function generating a vector of random numbers with mean zero and variance 1. Default is rnorm().}

\item{M}{Integer positiv real number. Square root of Gaussian kernels used for generation of the noise.}

\item{bdx}{Numeric postiv real number. Possibility to center the Gaussian kernels outside the range of x. Bdx simply extends the range by +-bdx and then a equidistant grid having M points from the interval range(x)-+bdx is formed giving the means for the Gaussian kernel bumps.}

\item{Fnorm}{Array containing the scaling to pointwise unit variance, if it is NULL this is done internally. This option is only included to speed up simulations.}
}
\value{
Array containing the realisations of the random field. Last dimension indexes the realisations.
}
\description{
Creates a sample from an 2D random field, which is the sum of 2-D Gaussian densities..
}
