% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Residuals.R
\name{get_moments}
\alias{get_moments}
\title{Computes the non-central moments of a functional data sets}
\usage{
get_moments(Y, moments)
}
\arguments{
\item{Y}{matrix containing the data. Last dimension enumerates the samples.}

\item{moments}{string of the form c("mu1", ..., "muk") containing the names.}

\item{cmoments}{boolean indicating whether an estimate of the cvariance matrix
of the moment vector is included in the output}

\item{squeeze}{boolean}
}
\value{
an list containing the moments and the corresponding moment residuals.
#' @export
}
\description{
Computes the non-central moments of a functional data sets
}
