% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GaussianKinematicFormulas.R
\name{GKFquantileApprox}
\alias{GKFquantileApprox}
\title{Approximates the upper tail probabilities of different fields using the corresponding Gaussian
kinematic formulas.
It approximates the tail probability of the absolute value by twice the
usual tail probability of the process.}
\usage{
GKFquantileApprox(alpha = 0.05, LKC, field = "t", df)
}
\arguments{
\item{alpha}{Numeric upper tail probabiltiy, which needs to be approximated.}

\item{LKC}{Vector estimated LKCs of the UGRFs of the field.}

\item{field}{String type of field.}

\item{df}{degrees of freedom.}
}
\value{
Numeric the approximated 1-alpha quantile.
}
\description{
Approximates the upper tail probabilities of different fields using the corresponding Gaussian
kinematic formulas.
It approximates the tail probability of the absolute value by twice the
usual tail probability of the process.
}
