% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoPEFunctions.R
\name{hatUc}
\alias{hatUc}
\title{This functions computes an upper confidence excursion set}
\usage{
hatUc(c, hatmu, hatsigma, tN, q, mu = NULL, inclusion = "inequal")
}
\arguments{
\item{c}{vector with the same length as hatmu containing the thresholding function}

\item{hatmu}{function to be thresholded to estimated the set mu1_C.
Usually it is an estimator of the target function mu}

\item{hatsigma}{scaling used to estimate the set mu1_C. Usually it is
the pointwise standard deviation og hatmu.}

\item{tN}{is inverse of the rate of the fCLT, i.e., usually 1/sqrt(N)}

\item{q}{quantile for the CoPE set}

\item{inclusion}{if "inequal" the upper excursion set is open, if "equal" the upper
excursion set is closed.}
}
\value{
if is.null(mu) a boolean vector indicating which positions are
in the upper confidence excursion set. Otherwise a list with elements
 \itemize{
  \item hatUc boolean vector indicating which positions are
in the upper confidence excursion set
  \item contain boolean indicating whether the upper confidence set is contained in
  the true upper excursion set
}
}
\description{
This functions computes an upper confidence excursion set
}
