% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomFieldGeneration.R
\name{HermiteSumNoise}
\alias{HermiteSumNoise}
\title{Creates sample paths from a 1D random field generated as a random sum of the first five Hermite polynomials.}
\usage{
HermiteSumNoise(N, x = seq(0, 1, length.out = 100), sigma = function(x) {   
   1 }, randNumber = rnorm)
}
\arguments{
\item{N}{Integer amount of realisations of the random field to be generated.}

\item{x}{Vector locations at which the random field is evaluated.}

\item{sigma}{Function computing the pointwise variance of the field. Default value is unit variance everywhere.}

\item{randNumber}{Function generating a vector of random numbers with mean zero and variance 1. Default is rnorm().}
}
\value{
Matrix containing the realisations of the random field as columns. The rows are the locations.
}
\description{
Creates sample paths from a 1D random field generated as a random sum of the first five Hermite polynomials.
}
