% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InsignificanceAnalysis.R
\name{IVobs}
\alias{IVobs}
\title{This functions computes some possible insignificance values.}
\usage{
IVobs(scopes, method)
}
\arguments{
\item{method}{if CoPE sets are computed either "selection" or
"extraction"(Default ).
 Otherwise if it is estimated for a statistical test
 using CoPE sets the options are "lrelevance", "relevance",
 "lequivalence", "equivalence"}

\item{SCoPEStype}{vector of same length as mu giving the possible alternative}
}
\value{
quantile of the maximum of the random variables
}
\description{
This functions computes some possible insignificance values.
}
