% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InsignificanceAnalysis.R
\name{IVs_IID}
\alias{IVs_IID}
\title{This functions computes the observational insignificance value (p-value )
for an IID Gaussian model the maximum of combinations of statistics of the form
-X[i], X[i] or |X[i]| where for {1,...,I} the random variables X[i] are
independently standard Gaussian distribtued.}
\usage{
IVs_IID(
  scopes,
  cdf = pnorm,
  cdf_abs = VGAM::pfoldnorm,
  rcdf = rnorm,
  Msim = 5e+05,
  KK = 5,
  ...
)
}
\arguments{
\item{cdf}{a cummulative distribution function taking a quantile q as input and
computing the probability that the random variable belonging to the cdf stays
below q}

\item{...}{further input variables for cdf}

\item{SCoPEStype}{vector of same length as mu giving the possible alternative}
}
\value{
quantile of the maximum of the random variables
}
\description{
This functions computes the observational insignificance value (p-value )
for an IID Gaussian model the maximum of combinations of statistics of the form
-X[i], X[i] or |X[i]| where for {1,...,I} the random variables X[i] are
independently standard Gaussian distribtued.
}
