% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Miscellaneous.R
\name{Ksquare_Gaussian}
\alias{Ksquare_Gaussian}
\title{This functions computes the values of Ksquare transformation under Gaussianity}
\usage{
Ksquare_Gaussian(N)
}
\arguments{
\item{N}{Vector grid for evaluation of the smoothed field}

\item{transformation}{The considered transformation of the moments}
}
\value{
Standard error under the assumption the data is Gaussian
}
\description{
This functions computes the values of Ksquare transformation under Gaussianity
}
