% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LKCEstimators.R
\name{LKC_estim_direct}
\alias{LKC_estim_direct}
\title{Estimates the Lipschitz killing curvatures of a Gaussian field over a rectangular domain using a finite difference estimator for the differential process. It is numerical (tested up to d=2) identical to the warping estimator proposed in  TaylorWorsley (2007).}
\usage{
LKC_estim_direct(R, subdiv = dim(R)[length(dim(R))])
}
\arguments{
\item{R}{Array (dimension K_1 x ... x K_d x N containing N-realizations of a standardaized Gaussian random field over a d-dimensional domain.)}

\item{subdiv}{Vector (containing the breakpoints to split the array in independent fields, e.g. for 2 sample tests)}
}
\value{
LKC Vector containing the Lipschitz killing curvatures of dimension greater than 1. Note that the 0-th LKC must be calculated seperately as the Euler characteristic of the domain.
}
\description{
IMPORTANT NOTES: the process must have mean 0 and variance 1 and current implementation only works for 1d and 2d domains of the random field.
}
