% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParametricQuantiles.R
\name{maxGauss_p}
\alias{maxGauss_p}
\title{This functions computes the probability to be below q for the maximum of
combinations of statistics of the form -X[i], X[i] or |X[i]|  where for
{1,...,I} the random variables X[i] are independently standard Gaussian
distribtued.}
\usage{
maxGauss_p(q, muC)
}
\arguments{
\item{muC}{list output from PreimageC}

\item{p}{numeric probability for which the quantile should be computed.}
}
\value{
quantile of the maximum of the random variables
}
\description{
This functions computes the probability to be below q for the maximum of
combinations of statistics of the form -X[i], X[i] or |X[i]|  where for
{1,...,I} the random variables X[i] are independently standard Gaussian
distribtued.
}
