% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParametricQuantiles.R
\name{maxGauss_quantile}
\alias{maxGauss_quantile}
\title{This functions computes the quatiles of the maximum of combinations of
statistics of the form -X[i], X[i] or |X[i]|  where for {1,...,I} the
random variables X[i] are independently standard Gaussian distribtued.}
\usage{
maxGauss_quantile(p, muC)
}
\arguments{
\item{p}{numeric probability for which the quantile should be computed.}

\item{muC}{list output from PreimageC}
}
\value{
quantile of the maximum of the random variables
}
\description{
This functions computes the quatiles of the maximum of combinations of
statistics of the form -X[i], X[i] or |X[i]|  where for {1,...,I} the
random variables X[i] are independently standard Gaussian distribtued.
}
