% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConfidenceBands.R
\name{maxQuantile}
\alias{maxQuantile}
\title{Computes the quantile of the maximum statistic}
\usage{
maxQuantile(R, alpha, method, coords, mask)
}
\arguments{
\item{R}{array of dimension K_1 x ... x K_D x N containing N-realizations of
a random field over a D-dimensional domain.}

\item{alpha}{numeric the upper tail probability}

\item{method}{}
}
\value{
quantile of the maximum process
}
\description{
Computes the quantile of the maximum statistic
}
