% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParametricQuantiles.R
\name{maxT_p}
\alias{maxT_p}
\title{This functions computes the quatiles of the maximum of combinations of
statistics of the form -X[i], X[i] or |X[i]|  where for {1,...,I} the
random variables X[i] are independently Student's t-distribtued with df
degrees of freedom}
\usage{
maxT_p(q, muC, df)
}
\arguments{
\item{muC}{list output from PreimageC}

\item{df}{numeric an integer giving the degrees of freedom}

\item{p}{numeric probability for which the quantile should be computed.}
}
\value{
quantile of the maximum of the random variables
}
\description{
This functions computes the quatiles of the maximum of combinations of
statistics of the form -X[i], X[i] or |X[i]|  where for {1,...,I} the
random variables X[i] are independently Student's t-distribtued with df
degrees of freedom
}
