% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Miscellaneous.R
\name{mean_Gaussian}
\alias{mean_Gaussian}
\title{This functions computes the true standard error depending on N for different
functionals of Gaussians}
\usage{
mean_Gaussian(transformation, N, hatd = NULL)
}
\arguments{
\item{transformation}{The considered transformation of the moments}

\item{N}{Vector grid for evaluation of the smoothed field}
}
\value{
Standard error under the assumption the data is Gaussian
}
\description{
This functions computes the true standard error depending on N for different
functionals of Gaussians
}
