% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Miscellaneous.R
\name{method_gen}
\alias{method_gen}
\title{Computes the scale field for a sample of 1d functional data}
\usage{
method_gen(name, SCoPEStype, mu1est, N, kN, ...)
}
\arguments{
\item{Y}{Matrix containing the data. Last dimension must indicate different samples.}

\item{xeval}{Vector grid for evaluation of the smoothed field}

\item{h}{Vector grid for the bandwidths for local linear smoothing or spacing between the knots in B-spline smoothing}

\item{method}{String Either 'loclin' for local linear smoothing or 'smoothingspline' for Bspline smoothing}

\item{kernel}{Function the kernel used for smoothing the data}

\item{Weights}{Matrix containing a smoothing matrix. If specified all other parameters for smoothing are ignored.}
}
\value{
Scale field
}
\description{
Computes the scale field for a sample of 1d functional data
}
