% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BootstrapQuantileApproximations.R
\name{MultiplierBootstrap}
\alias{MultiplierBootstrap}
\title{Multiplier bootstrap estimator for the quantile of the maximum of a random field. The function contains a multiplier-t version based on Telschow Schwartzmann (2018) "Simultaneous confidence bands for non-parametric regression with functional data" and a regular version based on Cherno. For large sample sizes the versions do agree, however, for small sample sizes the bootstrap-t has better covering rates, but a slightly higher variability in the estimate of the quantile than the simple version based on Cherno.}
\usage{
MultiplierBootstrap(R, params = NULL)
}
\arguments{
\item{R}{Array the last dimension enumerates the realizations of an residual field.}

\item{params}{list containing elements
\itemize{
  \item Mboots Numeric the amount of bootstrap replicates. Default is 5e3.
  \item alpha Numeric the targeted upper quantile of the maxiumum of the absolute value of the random field. Default is 0.95.
  \item method String the bootstrap version. Currently the options are "t" for the bootstrap-t and "regular". Default is "t".
}}
}
\value{
list with elements
 \itemize{
  \item z Vector containing the realizations of the bootstrap of the approximation of the maximum of the random field
  \item q alpha-quantile value of z
}
}
\description{
Multiplier bootstrap estimator for the quantile of the maximum of a random field. The function contains a multiplier-t version based on Telschow Schwartzmann (2018) "Simultaneous confidence bands for non-parametric regression with functional data" and a regular version based on Cherno. For large sample sizes the versions do agree, however, for small sample sizes the bootstrap-t has better covering rates, but a slightly higher variability in the estimate of the quantile than the simple version based on Cherno.
}
