% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BootstrapQuantileApproximations.R
\name{MultiplierBootstrapSplit}
\alias{MultiplierBootstrapSplit}
\title{Multiplier bootstrap estimator for the quantile of the maximum of a random field. The
function contains a multiplier-t version based on Telschow Schwartzmann (2019)
"Simultaneous confidence bands for non-parametric regression with functional data"
and a regular version based on Cherno. For large sample sizes the versions do agree,
however, for small sample sizes the bootstrap-t has better covering rates, but a slightly
higher variability in the estimate of the quantile than the simple version based on Cherno.}
\usage{
MultiplierBootstrapSplit(
  R,
  plus,
  minus,
  Q = NULL,
  alpha = 0.05,
  Mboots = 5000,
  method = "t",
  weights = "rademacher"
)
}
\arguments{
\item{R}{array of dimension K_1 x ... x K_D x N containing N-realizations of
a random field over a D-dimensional domain.}

\item{Q}{array of dimension K_1 x ... x K_D x N containing N-realizations of
a random field over a D-dimensional domain. Default NULL, i.e. one sample
case.}

\item{alpha}{numeric the targeted upper quantile of the maxiumum of the
absolute value of the random field. Default is 0.95.}

\item{Mboots}{numeric amount of bootstrap replicates. Default is 5e3.}

\item{method}{string specifies the bootstrap version. Options are "t" and
"regular". Default is "t".}

\item{weights}{string specifying the multipliers. Options are "gaussian",
"rademacher" and "mammen". Default is "rademacher".}
}
\value{
list with elements
 \itemize{
  \item z Vector containing the realizations of the bootstrap of the approximation of the maximum of the random field
  \item q alpha-quantile value of z
}
}
\description{
Multiplier bootstrap estimator for the quantile of the maximum of a random field. The
function contains a multiplier-t version based on Telschow Schwartzmann (2019)
"Simultaneous confidence bands for non-parametric regression with functional data"
and a regular version based on Cherno. For large sample sizes the versions do agree,
however, for small sample sizes the bootstrap-t has better covering rates, but a slightly
higher variability in the estimate of the quantile than the simple version based on Cherno.
}
