% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BootstrapQuantileApproximations.R
\name{NonParametricBootstrap}
\alias{NonParametricBootstrap}
\title{Non-parametric bootstrap estimator for the quantile of the maximum of a random field. The function contains a bootstrap-t version based on Degras (2011) "Simultaneous confidence bands for non-parametric regression with functional data" and a simple non-parametric bootstrap. For large sample sizes the versions do agree, however, for small sample sizes the bootstrap-t has better covering rates, but higher variability in the estimate of the quantile than the simple non-paramtric bootstrap.}
\usage{
NonParametricBootstrap(A, params = NULL)
}
\arguments{
\item{A}{Array the last dimension enumerates the realizations of the random field.}

\item{params}{list containing elements
\itemize{
  \item Mboots Numeric the amount of bootstrap replicates. Default is 5e3.
  \item alpha Numeric the targeted upper quantile of the maxiumum of the absolute value of the random field. Default is 0.95.
  \item method String the bootstrap version. Currently the options are "t" for the bootstrap-t and "regular". Default is "t".
}}
}
\value{
list with elements
 \itemize{
  \item z Vector containing the realizations of the bootstrap of the approximation of the maximum of the random field
  \item q alpha-quantile value of z
}
}
\description{
Non-parametric bootstrap estimator for the quantile of the maximum of a random field. The function contains a bootstrap-t version based on Degras (2011) "Simultaneous confidence bands for non-parametric regression with functional data" and a simple non-parametric bootstrap. For large sample sizes the versions do agree, however, for small sample sizes the bootstrap-t has better covering rates, but higher variability in the estimate of the quantile than the simple non-paramtric bootstrap.
}
