% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomFieldGeneration.R
\name{OUNoise}
\alias{OUNoise}
\title{Creates sample paths from a 1D Ornstein Uhlenbeck process.}
\usage{
OUNoise(N, x = seq(0, 1, length.out = 100), sigma = function(x) {     1 },
  alphaOU = 5, sigmaOU = sqrt(10), gamma = NULL)
}
\arguments{
\item{N}{Integer amount of realisations of the random field to be generated.}

\item{x}{Vector locations at which the random field is evaluated.}

\item{sigma}{Function computing the pointwise variance of the field. Default value is unit variance everywhere.}

\item{alphaOU}{Numeric value of the retraction parameter in an OU process. Default is 5.}

\item{sigmaOU}{Numeric value of the sigma parameter in an OU process. Default is sqrt(10).}

\item{ganna}{Vector of the same length as x giving the mean of the OU process.}
}
\value{
Matrix containing the realisations of the random field as columns. The rows are the locations.
}
\description{
Creates sample paths from a 1D Ornstein Uhlenbeck process.
}
