% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BootstrapQuantileApproximations.R
\name{ParametricBootstrap}
\alias{ParametricBootstrap}
\title{Parametric bootstrap estimator for the quantile of the maximum of a random field.
The function contains a bootstrap-t version based on Degras (2011)
"Simultaneous confidence bands for non-parametric regression with functional data" and
a simple non-parametric bootstrap. For large sample sizes the versions do agree, however,
for small sample sizes the bootstrap-t has better covering rates, but higher variability
in the estimate of the quantile than the simple non-paramtric bootstrap.}
\usage{
ParametricBootstrap(A, alpha = 0.05, Mboots = 5000, method = "t")
}
\arguments{
\item{A}{array of dimension K_1 x ... x K_D x N containing N-realizations of
a random field over a D-dimensional domain.}

\item{alpha}{numeric the targeted upper quantile of the maxiumum of the
absolute value of the random field. Default is 0.95.}

\item{Mboots}{numeric amount of bootstrap replicates. Default is 5e3.}

\item{method}{string specifies the bootstrap version. Options are "t" and
"regular". Default is "t".}
}
\value{
list with elements
 \itemize{
  \item z Vector containing the realizations of the bootstrap of the approximation of the maximum of the random field
  \item q alpha-quantile value of z
}
}
\description{
Parametric bootstrap estimator for the quantile of the maximum of a random field.
The function contains a bootstrap-t version based on Degras (2011)
"Simultaneous confidence bands for non-parametric regression with functional data" and
a simple non-parametric bootstrap. For large sample sizes the versions do agree, however,
for small sample sizes the bootstrap-t has better covering rates, but higher variability
in the estimate of the quantile than the simple non-paramtric bootstrap.
}
