% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFunctions.R
\name{plot_covSim}
\alias{plot_covSim}
\title{This functions computes the standard}
\usage{
plot_covSim(covSim, ylims = c(0.3, 1), title = "", legend.position = "none")
}
\arguments{
\item{title}{Vector grid for evaluation of the smoothed field}

\item{x}{The considered transformation of the moments}

\item{scb}{Vector grid for evaluation of the smoothed field}

\item{scb2}{Vector grid for evaluation of the smoothed field}

\item{true}{Vector grid for evaluation of the smoothed field}
}
\value{
Standard error under the assumption the data is Gaussian
}
\description{
This functions computes the standard
}
