% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFunctions.R
\name{plot_scb}
\alias{plot_scb}
\title{This functions computes the standard}
\usage{
plot_scb(
  x,
  scb,
  scb2 = NULL,
  true = NULL,
  s1name = "est Sample1",
  s2name = "est Sample2",
  title = "",
  legend.position = "none"
)
}
\arguments{
\item{x}{The considered transformation of the moments}

\item{scb}{Vector grid for evaluation of the smoothed field}

\item{scb2}{Vector grid for evaluation of the smoothed field}

\item{true}{Vector grid for evaluation of the smoothed field}

\item{title}{Vector grid for evaluation of the smoothed field}
}
\value{
Standard error under the assumption the data is Gaussian
}
\description{
This functions computes the standard
}
