% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CriticalSets.R
\name{PreimageC}
\alias{PreimageC}
\title{Estimates the preimage of the set indicated by the matrix C using
the thickening estimator proposed in Telschow et al (2022)
"SCoPES: A versatile framework of simultaneous inference".}
\usage{
PreimageC(C, hatmu, hatsigma, tN, kN = log(tN^-2)/5, method = "extraction")
}
\arguments{
\item{C}{T x nlvl matrix containing nlvl functions for which the preimage
should be computed. At the moment only the first and last function
matter, i.e., they define a tube.}

\item{hatmu}{function to be thresholded to estimated the set mu1_C.
Usually it is an estimator of the target function mu}

\item{hatsigma}{scaling used to estimate the set mu1_C. Usually it is
the pointwise standard deviation og hatmu.}

\item{tN}{is inverse of the rate of the fCLT, i.e., usually 1/sqrt(N)}

\item{kN}{is the thickening of the set (Default log(tN^-2) / 5)}

\item{method}{if CoPE sets are computed either "selection" or
"extraction"(Default ).
 Otherwise if it is estimated for a statistical test
 using CoPE sets the options are "lrelevance", "relevance",
 "lequivalence", "equivalence"}
}
\value{
list with elements
 \itemize{
  \item minus boolean vector indicating which locations are in the set
        mu1_C^-
  \item plus boolean vector indicating which locations are in the set
        mu1_C^+
}
}
\description{
Estimates the preimage of the set indicated by the matrix C using
the thickening estimator proposed in Telschow et al (2022)
"SCoPES: A versatile framework of simultaneous inference".
}
