% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Miscellaneous.R
\name{residualsSNR}
\alias{residualsSNR}
\title{Computes the SNR residuals for a sample of 1d functional data}
\usage{
residualsSNR(Y, bias = TRUE)
}
\arguments{
\item{Y}{Matrix containing the data. Last dimension must indicate different samples.}

\item{bias}{Boolean TRUE means variance estimated using 1/N, FALSE estimates it with 1/(N-1)}
}
\value{
Scale field
}
\description{
Computes the SNR residuals for a sample of 1d functional data
}
