% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Miscellaneous.R
\name{scaleField}
\alias{scaleField}
\title{Computes the scale field for a sample of 1d functional data}
\usage{
scaleField(Y, x = seq(0, 1, length.out = nrow(Y)), xeval = seq(0, 1,
  length.out = 2 * nrow(Y)), h = seq(x[2], 0.1, length.out = 20),
  method = "loclin", kernel = locpol::gaussK, Weights = NULL)
}
\arguments{
\item{Y}{Matrix containing the data. Last dimension must indicate different samples.}

\item{xeval}{Vector grid for evaluation of the smoothed field}

\item{h}{Vector grid for the bandwidths for local linear smoothing or spacing between the knots in B-spline smoothing}

\item{method}{String Either 'loclin' for local linear smoothing or 'smoothingspline' for Bspline smoothing}

\item{kernel}{Function the kernel used for smoothing the data}

\item{Weights}{Matrix containing a smoothing matrix. If specified all other parameters for smoothing are ignored.}
}
\value{
Scale field
}
\description{
Computes the scale field for a sample of 1d functional data
}
