% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConfidenceBands.R
\name{scb_mean}
\alias{scb_mean}
\title{Computes simultaneous confidence bands for the mean of a sample from a one dimensional functional signal plus noise model. It is possible to choose between different estimators for the quantile.}
\usage{
scb_mean(Y, level = 0.95, method = "tGKF", param_method = NULL)
}
\arguments{
\item{Y}{Array of dimension K_1 x ... x K_d x N containing N-realizations of a Gaussian random field over a d-dimensional domain.}

\item{level}{Numeric the targeted covering probability. Must be strictly between 0 and 1.}

\item{method}{Stringspecifying the method to construt the scb, i.e. estimatate the quantile. Current options are "tGKF", "GKF", "Bootstrap", "Bootstrapt", "MultiplierBootstrap", "MultiplierBootstrapt". Default value is "tGKF".}

\item{param_method}{list containing the parameters for 'method'. The list must contain the following elements, otherwise default values are set:
 \itemize{
  \item For method either "tGKF" or "GKF": L0 (integer) the Euler characteristic of the domain of the random fields Y [default is 1], LKC_estim (function) a function estimating the Lipschitz-Killing curvatures from the normalized residuals [default is LKC_estim_direct].
  \item For method "Bootstrap", "Bootstrapt", "MultiplierBootstrap", "MultiplierBootstrapt": Mboots (positiv integer) the amount of bootstrap replicates [default 5e3].
}}
}
\value{
list with elements
 \itemize{
  \item hatmean pointwise sample mean
  \item scb list containing the upper and lower bounds of the simultaneous confidence band
  \item level targeted covering probability
  \item q quantile of the maximum of the residual field
}
}
\description{
Computes simultaneous confidence bands for the mean of a sample from a one dimensional functional signal plus noise model. It is possible to choose between different estimators for the quantile.
}
