% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConfidenceBands.R
\name{scb_meandiff}
\alias{scb_meandiff}
\title{Computes simultaneous confidence bands for the difference of the means of two samples from one dimensional functional signal plus noise models. It is possible to choose between different estimators for the quantile.}
\usage{
scb_meandiff(Y1, Y2, level = 0.95, method = "tGKF",
  param_method = NULL)
}
\arguments{
\item{Y1}{Array dimension K_1 x ... x K_d x N containing N-realizations of a Gaussian random field over a d-dimensional domain. This is sample 1.}

\item{Y2}{Array dimension K_1 x ... x K_d x N containing N-realizations of a Gaussian random field over a d-dimensional domain. This is sample 2.}

\item{level}{Numeric containing the level of confidence for the constructed SCBs.}

\item{method}{String the name of the method used for quantile estimation. Possibilities are "tGKF" and "GKF". Default value is "tGKF".}
}
\value{
LKC Vector containing the Lipschitz killing curvatures of dimension greater than 1. Note that the 0-th LKC must be calculated seperately as the Euler characteristic of the domain.
}
\description{
Computes simultaneous confidence bands for the difference of the means of two samples from one dimensional functional signal plus noise models. It is possible to choose between different estimators for the quantile.
}
