% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConfidenceBands.R
\name{SCBglm}
\alias{SCBglm}
\title{Computes simultaneous confidence bands for a contrast in a functional linear model. It is possible to choose between different estimators for the quantile.}
\usage{
SCBglm(x = seq(0, 1, length.out = 100), y, X = NULL, c = c(1, -1),
  xlim = c(0, 1), level = 0.95, quantile.method = "tGKF", Mboots = 1000,
  evalN = ifelse(is.vector(x), 4 * length(x), 4 * length(x[1, ])),
  LinWeights = NULL, bw = NULL, kernel = locpol::gaussK,
  sigma.boots = "boots", quantile = NULL)
}
\arguments{
\item{X}{Matrix (dimensions N x P. It is the design matrix of the pointwise functional linear model.)}

\item{c}{Vector (of length P defining the contrast of interest.)}

\item{level}{Numeric (the targeted covering probability. Must be strictly between 0 and 1.)}

\item{Mboots}{Numeric (the number of bootstrap replicats used, if the quantile is estimated using a bootstrap method. Default value is 5000.)}

\item{Y}{Array (dimension K_1 x ... x K_d x N containing N-realizations of a Gaussian random field over a d-dimensional domain.)}

\item{method}{String (the name of the method used for quantile estimation. Possibilities are "tGKF", "GKF", "Bootstrap", "Bootstrapt", "MultiplierBootstrap", "MultiplierBootstrapt". Default value is "tGKF".)}
}
\value{
list with elements
 \itemize{
  \item hatmean pointwise sample mean
  \item scb list containing the upper and lower bounds of the simultaneous confidence band
  \item level targeted covering probability
  \item q quantile of the maximum of the residual field
}
}
\description{
Computes simultaneous confidence bands for a contrast in a functional linear model. It is possible to choose between different estimators for the quantile.
}
