% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoPEFunctions.R
\name{SCoPES}
\alias{SCoPES}
\title{This functions computes the SCoPES corresponding to an estimator and a set
of functions given as a matrix with columns being the cut-off functions.}
\usage{
SCoPES(
  alpha,
  C,
  x,
  hatmu,
  hatsigma,
  tN,
  method = "extraction",
  R = NULL,
  I = NULL,
  inclusion = list(L = "inequal", U = "inequal"),
  mu = NULL
)
}
\arguments{
\item{alpha}{numeric between 0 and 1. It will produce (1-alpha) SCoPES.}

\item{C}{T x nlvl matrix containing nlvl functions for which the preimage
should be computed. At the moment only the first and last function
matter, i.e., they define a tube.}

\item{hatmu}{function to be thresholded to estimated the set mu1_C.
Usually it is an estimator of the target function mu}

\item{hatsigma}{scaling used to estimate the set mu1_C. Usually it is
the pointwise standard deviation og hatmu.}

\item{tN}{is inverse of the rate of the fCLT, i.e., usually 1/sqrt(N)}

\item{method}{this can be either a string specifying the method, i.e.,
"selection", "extraction"(Default ), lrelevance", "relevance",
"lequivalence", "equivalence". The latter for are the hypothesis tests
described in Telschow et al 2022. In this case the quantile q is obtained
by a bootstrap on the thickened set.
It can also be a method specifying the estimation of the quantile $q$.}

\item{R}{array containing as columns residuals which can be used to bootstrap
the quantile q. This input is mandatory if "method" is only a string.}

\item{inclusion}{list with elements
 \itemize{
  \item L string either "inequal" or "equal" determining whether the lower
  excursion set uses < c or <=c.
  \item U string either "inequal" or "equal" determining whether the upper
  excursion set uses > c or >=c.
}}
}
\value{
Standard error under the assumption the data is Gaussian
}
\description{
This functions computes the SCoPES corresponding to an estimator and a set
of functions given as a matrix with columns being the cut-off functions.
}
