% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomFieldGeneration.R
\name{SinCosSumNoise}
\alias{SinCosSumNoise}
\title{Creates sample paths from a 1D Gaussian fields generated as a random sum of sine and cosine functions with frequency pi/2 and standard Gaussians as coefficients.}
\usage{
SinCosSumNoise(N, x = seq(0, 1, length.out = 100), sigma = function(x)
  {     1 })
}
\arguments{
\item{N}{Integer amount of realisations of the random field to be generated.}

\item{x}{Vector locations at which the random field is evaluated.}

\item{sigma}{Vector specifies the pointwise variance of the field. Default value is unit variance everywhere.}
}
\value{
Matrix containing the realisations of the random field as columns. The rows are the locations.
}
\description{
Creates sample paths from a 1D Gaussian fields generated as a random sum of sine and cosine functions with frequency pi/2 and standard Gaussians as coefficients.
}
