% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomFieldGeneration.R
\name{SquaredExp2DNoise}
\alias{SquaredExp2DNoise}
\title{Creates a sample from an Gaussian isotropic 2D random field having a square exponential function: C(h)=exp( -h^2/(4*nu^2) ) as covariance structure by smoothing independent noise with a Gaussian kernel.}
\usage{
SquaredExp2DNoise(N, x = seq(1, 50, by = 1), sigma = function(x) {    
  matrix(1, length(x), length(x)) }, nu = 5)
}
\arguments{
\item{N}{Integer amount of realisations of the random field to be generated.}

\item{x}{Vector locations at which the random field is evaluated.}

\item{sigma}{Function computing the pointwise variance of the field. Default value is unit variance everywhere.}

\item{nu}{Numeric positiv real number. Smoothing bandwidth of the Gaussian kernel}
}
\value{
Array containing the realisations of the random field. Last dimension indexes the realisations.
}
\description{
Creates a sample from an Gaussian isotropic 2D random field having a square exponential function: C(h)=exp( -h^2/(4*nu^2) ) as covariance structure by smoothing independent noise with a Gaussian kernel.
}
