% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Miscellaneous.R
\name{WF_transform}
\alias{WF_transform}
\title{This functions computes the Wilson & Hilferty transform}
\usage{
WF_transform(x, df = 2)
}
\arguments{
\item{transformation}{The considered transformation of the moments}

\item{N}{Vector grid for evaluation of the smoothed field}
}
\value{
Standard error under the assumption the data is Gaussian
}
\description{
This functions computes the Wilson & Hilferty transform
}
