% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Miscellaneous.R
\name{Z2_transform}
\alias{Z2_transform}
\title{This functions computes the Wilson & Hilferty transform}
\usage{
Z2_transform(g2, N)
}
\arguments{
\item{N}{Vector grid for evaluation of the smoothed field}

\item{transformation}{The considered transformation of the moments}
}
\value{
Standard error under the assumption the data is Gaussian
}
\description{
This functions computes the Wilson & Hilferty transform
}
